Clear optimize

+++ app/src/main/java/one/mixin/android/job/MessageDeleteJob.kt
@@ -3,13 +3,13 @@ package one.mixin.android.job
 import com.birbit.android.jobqueue.Params
 import kotlinx.coroutines.runBlocking
 import one.mixin.android.Constants.DB_DELETE_LIMIT
-import one.mixin.android.db.deleteMessageByConversationId
+import one.mixin.android.db.deleteMessageByIds
 import one.mixin.android.util.chat.InvalidateFlow
 
 class MessageDeleteJob(
     private val conversationId: String,
-    private val deleteMention: Boolean = false,
-    private val deleteConversation: Boolean = false
+    private val lastRowId: Long,
+    private val deleteConversation:Boolean
 ) :
     BaseJob(Params(PRIORITY_UI_HIGH).addTags(GROUP).groupBy("message_delete").persist()) {
 
@@ -17,30 +17,22 @@ class MessageDeleteJob(
 
     companion object {
         const val GROUP = "MessageDeleteJob"
-        private const val serialVersionUID = 1L
+        private const val serialVersionUID = 2L
     }
 
     override fun onRun() = runBlocking {
-        if (deleteMention) {
-            val deleteTimes =
-                messageMentionDao.countDeleteMessageByConversationId(conversationId) / DB_DELETE_LIMIT + 1
-            repeat(deleteTimes) {
-                messageMentionDao.deleteMessageByConversationId(conversationId, DB_DELETE_LIMIT)
-            }
-        } else {
-            val deleteTimes =
-                messageDao.countDeleteMessageByConversationId(conversationId) / DB_DELETE_LIMIT + 1
-            repeat(deleteTimes) {
-                messageFts4Dao.deleteByMessageIds(
-                    messageDao.getMessageIdsByConversationId(
-                        conversationId,
-                        DB_DELETE_LIMIT
-                    )
-                )
-                appDatabase.deleteMessageByConversationId(conversationId, DB_DELETE_LIMIT)
-            }
+        val deleteTimes =
+            messageDao.countDeleteMessageByConversationId(conversationId) / DB_DELETE_LIMIT + 1
+        repeat(deleteTimes) {
+            val ids = messageDao.getMessageIdsByConversationId(
+                conversationId,
+                lastRowId,
+                DB_DELETE_LIMIT
+            )
+            jobManager.addJobInBackground(MessageFtsDeleteJob(ids))
+            appDatabase.deleteMessageByIds(ids)
         }
-        if (deleteConversation) {
+        if (deleteConversation && messageDao.findLastMessageRowId(conversationId) == lastRowId) {
             conversationDao.deleteConversationById(conversationId)
         }
         InvalidateFlow.emit(conversationId)
